Separated test funcs into different files
